+++ src/systems/a320_systems/src/pneumatic.rs
@@ -3,11 +3,11 @@ use std::{f64::consts::PI, time::Duration};
 
 use uom::si::{
     f64::*,
+    mass_rate::kilogram_per_second,
     pressure::psi,
     ratio::ratio,
     thermodynamic_temperature::degree_celsius,
     volume::{cubic_meter, gallon},
-    volume_rate::cubic_meter_per_second,
 };
 
 use systems::{
@@ -19,7 +19,7 @@ use systems::{
         CrossBleedValveSelectorKnob, CrossBleedValveSelectorMode,
         EngineCompressionChamberController, EngineModeSelector, EngineState, PneumaticContainer,
         PneumaticPipe, PneumaticValveSignal, Precooler, PressurisedReservoirWithExhaustValve,
-        PressurizeableReservoir, TargetPressureSignal, VariableVolumeContainer,
+        PressurizeableReservoir, TargetPressureTemperatureSignal, VariableVolumeContainer,
     },
     shared::{
         pid::PidController, update_iterator::MaxStepLoop, ControllerSignal, ElectricalBusType,
@@ -214,7 +214,7 @@ impl A320Pneumatic {
         engines: [&(impl EngineCorrectedN1 + EngineCorrectedN2); 2],
         overhead_panel: &A320PneumaticOverheadPanel,
         engine_fire_push_buttons: &impl EngineFirePushButtons,
-        apu: &impl ControllerSignal<TargetPressureSignal>,
+        apu: &impl ControllerSignal<TargetPressureTemperatureSignal>,
     ) {
         self.physics_updater.update(context);
 
@@ -235,7 +235,7 @@ impl A320Pneumatic {
         engines: [&(impl EngineCorrectedN1 + EngineCorrectedN2); 2],
         overhead_panel: &A320PneumaticOverheadPanel,
         engine_fire_push_buttons: &impl EngineFirePushButtons,
-        apu: &impl ControllerSignal<TargetPressureSignal>,
+        apu: &impl ControllerSignal<TargetPressureTemperatureSignal>,
     ) {
         self.apu_compression_chamber.update(apu);
 
@@ -836,7 +836,7 @@ impl EngineBleedAirSystem {
             ),
             engine_starter_exhaust: PneumaticExhaust::new(3e-2, 3e-2, Pressure::new::<psi>(0.)),
             engine_starter_valve: DefaultValve::new_closed(),
-            precooler: Precooler::new(5.),
+            precooler: Precooler::new(180. * 2.),
         }
     }
 
@@ -1030,8 +1030,21 @@ impl PneumaticContainer for EngineBleedAirSystem {
         self.precooler_outlet_pipe.temperature()
     }
 
-    fn change_fluid_amount(&mut self, volume: Volume) {
-        self.precooler_outlet_pipe.change_fluid_amount(volume)
+    fn mass(&self) -> Mass {
+        self.precooler_outlet_pipe.mass()
+    }
+
+    fn change_fluid_amount(
+        &mut self,
+        fluid_amount: Mass,
+        fluid_temperature: ThermodynamicTemperature,
+        fluid_pressure: Pressure,
+    ) {
+        self.precooler_outlet_pipe.change_fluid_amount(
+            fluid_amount,
+            fluid_temperature,
+            fluid_pressure,
+        )
     }
 
     fn update_temperature(&mut self, temperature: TemperatureInterval) {
@@ -1190,8 +1203,18 @@ impl PneumaticContainer for PackComplex {
         self.pack_container.temperature()
     }
 
-    fn change_fluid_amount(&mut self, volume: Volume) {
-        self.pack_container.change_fluid_amount(volume);
+    fn mass(&self) -> Mass {
+        self.pack_container.mass()
+    }
+
+    fn change_fluid_amount(
+        &mut self,
+        fluid_amount: Mass,
+        fluid_temperature: ThermodynamicTemperature,
+        fluid_pressure: Pressure,
+    ) {
+        self.pack_container
+            .change_fluid_amount(fluid_amount, fluid_temperature, fluid_pressure);
     }
 
     fn update_temperature(&mut self, temperature_change: TemperatureInterval) {
@@ -1223,15 +1246,15 @@ impl PackFlowValveController {
     fn new(context: &mut InitContext, engine_number: usize) -> Self {
         Self {
             pack_toggle_pb_id: context
-                .get_identifier(format!("AIRCOND_PACK{}_TOGGLE", engine_number)),
+                .get_identifier(format!("OVHD_COND_PACK_{}_PB_IS_ON", engine_number)),
             pack_pb_is_auto: true,
             pid: PidController::new(0., 0.05, 0., 0., 1., 0.75, 1.),
         }
     }
 
-    fn update(&mut self, context: &UpdateContext, pack_flow_valve_flow_rate: VolumeRate) {
+    fn update(&mut self, context: &UpdateContext, pack_flow_valve_flow_rate: MassRate) {
         self.pid.next_control_output(
-            pack_flow_valve_flow_rate.get::<cubic_meter_per_second>(),
+            pack_flow_valve_flow_rate.get::<kilogram_per_second>(),
             Some(context.delta()),
         );
     }
@@ -1331,7 +1354,7 @@ mod tests {
         pneumatic::{
             BleedMonitoringComputerChannelOperationMode, ControllablePneumaticValve,
             CrossBleedValveSelectorMode, EngineState, PneumaticContainer, PneumaticValveSignal,
-            TargetPressureSignal,
+            TargetPressureTemperatureSignal,
         },
         shared::{
             ApuBleedAirValveSignal, ControllerSignal, ElectricalBusType, ElectricalBuses,
@@ -1347,9 +1370,8 @@ mod tests {
     use std::{fs, fs::File, time::Duration};
 
     use uom::si::{
-        f64::*, length::foot, pressure::psi, ratio::ratio,
+        f64::*, length::foot, mass_rate::kilogram_per_second, pressure::psi, ratio::ratio,
         thermodynamic_temperature::degree_celsius, velocity::knot,
-        volume_rate::cubic_meter_per_second,
     };
 
     use super::{A320Pneumatic, A320PneumaticOverheadPanel};
@@ -1357,12 +1379,14 @@ mod tests {
     struct TestApu {
         bleed_air_valve_signal: ApuBleedAirValveSignal,
         bleed_air_pressure: Pressure,
+        bleed_air_temperature: ThermodynamicTemperature,
     }
     impl TestApu {
         fn new() -> Self {
             Self {
                 bleed_air_valve_signal: ApuBleedAirValveSignal::new_closed(),
                 bleed_air_pressure: Pressure::new::<psi>(14.7),
+                bleed_air_temperature: ThermodynamicTemperature::new::<degree_celsius>(15.),
             }
         }
 
@@ -1374,6 +1398,10 @@ mod tests {
             self.bleed_air_pressure = pressure;
         }
 
+        fn set_bleed_air_temperature(&mut self, temperature: ThermodynamicTemperature) {
+            self.bleed_air_temperature = temperature;
+        }
+
         fn set_bleed_air_valve_signal(&mut self, signal: ApuBleedAirValveSignal) {
             self.bleed_air_valve_signal = signal;
         }
@@ -1383,9 +1411,12 @@ mod tests {
             Some(self.bleed_air_valve_signal)
         }
     }
-    impl ControllerSignal<TargetPressureSignal> for TestApu {
-        fn signal(&self) -> Option<TargetPressureSignal> {
-            Some(TargetPressureSignal::new(self.bleed_air_pressure))
+    impl ControllerSignal<TargetPressureTemperatureSignal> for TestApu {
+        fn signal(&self) -> Option<TargetPressureTemperatureSignal> {
+            Some(TargetPressureTemperatureSignal::new(
+                self.bleed_air_pressure,
+                self.bleed_air_temperature,
+            ))
         }
     }
 
@@ -1807,6 +1838,12 @@ mod tests {
 
         fn set_bleed_air_running(mut self) -> Self {
             self.command(|a| a.apu.set_bleed_air_pressure(Pressure::new::<psi>(42.)));
+            self.command(|a| {
+                a.apu
+                    .set_bleed_air_temperature(ThermodynamicTemperature::new::<degree_celsius>(
+                        250.,
+                    ))
+            });
 
             self.set_apu_bleed_valve_signal(ApuBleedAirValveSignal::new_open())
                 .set_apu_bleed_air_pb(true)
@@ -1853,7 +1890,7 @@ mod tests {
         }
 
         fn set_pack_flow_pb_is_auto(mut self, number: usize, is_auto: bool) -> Self {
-            self.write_by_name(&format!("AIRCOND_PACK{}_TOGGLE", number), is_auto);
+            self.write_by_name(&format!("OVHD_COND_PACK_{}_PB_IS_ON", number), is_auto);
 
             self
         }
@@ -1905,7 +1942,7 @@ mod tests {
             self.query(|a| a.pneumatic.fadec.is_single_vs_dual_bleed_config())
         }
 
-        fn pack_flow_valve_flow(&self, engine_number: usize) -> VolumeRate {
+        fn pack_flow_valve_flow(&self, engine_number: usize) -> MassRate {
             self.query(|a| {
                 a.pneumatic.packs[engine_number - 1]
                     .pack_flow_valve
@@ -1946,8 +1983,8 @@ mod tests {
         Pressure::new::<psi>(0.5)
     }
 
-    fn flow_rate_tolerance() -> VolumeRate {
-        VolumeRate::new::<cubic_meter_per_second>(0.1)
+    fn flow_rate_tolerance() -> MassRate {
+        MassRate::new::<kilogram_per_second>(0.1)
     }
 
     // Just a way for me to plot some graphs